refactor: rename DynamicNodeIdConfig to NodeIdProvider
